refactor(connection.py, console_ui.py): Simplify connection handling and command processing
- Refactored P2PConnection class to inherit from multiple mixins, enhancing modularity and reducing code complexity.
- Updated ConsoleUI to delegate command handling to dedicated functions, improving readability and maintainability.
- Removed redundant imports and methods in ConsoleUI, streamlining the command processing logic.
